Fixes Heap-buffer-overflow READ in pcpp::cross_platform_memmem (#914)
* Fixes Heap-buffer-overflow READ in pcpp::cross_platform_memmem

Fixes Heap-buffer-overflow READ in pcpp::cross_platform_memmem:
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=43408

The `haystackLen` for `cross_platform_memmem` call in `SipRequestFirstLine::parseVersion` is calculated as `m_SipRequest->getDataLen() - m_UriOffset`. However when `m_DataLen` is 4 and `m_UriOffset` is 5 it results in -1 which is converted into max uint64. This leads to memory read out of data buffer bounds.

* Remove trailing tab

* Add the regression test