3rdParty: upgrade minimum cmake version to 3.12
Done to avoid warning in 3rdParty/LightPcapNg/CMakeLists.txt
and error in 3rdParty/MemPlumber/MemPlumber/CMakeLists.txt,
when using cmake version 4.0

CMake 3.12 seems to be the version used by other files in this project,
hence the upgrade to the particular version.

As a point of reference,
Debian 10/Buster was End Of Life in 2024.
Its cmake package is of version 3.13.
So the possibility for someone to get a compilation error from this change is nearing null.