Remove protected field `m_DeviceOpened` from `IDevice` interface.
The C++ core guidelines recommend against usages of protected fields, due to the maintainability issues in large class hierarchies.
As `IDevice` is the base interface of every device, it is better to let concrete implementations handle their own state detection than to force a boolean to be used everywhere.
A `PCapLiveDevice` for instance, can instead use the state of its `PcapHandle` for state detection.