Converted `parseNextLayer` to static dispatch. Added protected `doParseNextLayer` to do the actual parsing.
The change is necessary due to the way C++ function hiding works in derived classes.
If an overload `Derived::foo(int)` is declared, it hides all overloads `Base::foo(...)`. This is also true for virtual methods.

The way backwards compatibility with the old `parseNextLayer` signature is done, is that there exist 2 overloads of the method,
a parameter-less overload and one that takes `ParserConfiguration`. Given that all derived `Layer` classes implement
a override of `parseNextLayer(ParserConfiguration const&)`, the pattern hides the parameter less overload for all derived classes.
This is undesirable as it breaks backwards compatibility of the API.

The implemented solution is to split the function into two logical pieces, the API (`parseNextLayer`) and the implementation: (`doParseNextLayer`).
The API part are to be public facing static functions that pipe the input to `doParseNextLayer`, while the new method does the actual parsing.

An alternative solution to use `using Layer::parseNextLayer` on all derived classes was considered and discarded, due to lack of compiler enforcement.
An `doParseNextLayer` will always be enforced to exist, while the compiler can not be made to require `using Layer::parseNextLayer` to be included in derived classes.